+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/response/search/aggregation/metrics/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Protocol Buffer utilities for metrics aggregation responses.
+ * Contains converters from OpenSearch internal metrics aggregation results to Protocol Buffer messages.
+ * <p>
+ * Metrics aggregations compute metrics (like min, max, avg, sum, cardinality) over a set of documents.
+ *
+ * @see org.opensearch.search.aggregations.metrics
+ */
+package org.opensearch.transport.grpc.proto.response.search.aggregation.metrics;